backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
yt-dlp>=2024.3.10
ffmpeg-python>=0.2.0
aiofiles>=23.2.1
pydantic>=2.6.4
websockets>=12.0
python-multipart>=0.0.9
